+++ src/WPBT/WPBT_Extras.php
@@ -91,6 +91,19 @@ class WPBT_Extras {
 				'description' => esc_html__( 'Disable sending emails to the admin user for successful autoupdates. Only emails indicating failures of the autoupdate process are sent.', 'wordpress-beta-tester' ),
 			)
 		);
+
+		add_settings_field(
+			'remove_auto_installed_plugins',
+			null,
+			array( 'WPBT_Settings', 'checkbox_setting' ),
+			'wp_beta_tester_extras',
+			'wp_beta_tester_email',
+			array(
+				'id'          => 'remove_auto_installed_plugins',
+				'title'       => esc_html__( 'Delete auto-installed plugins.', 'wordpress-beta-tester' ),
+				'description' => esc_html__( 'Akismet and Hello Dolly are automatically installed with beta testing offers.', 'wordpress-beta-tester' ),
+			)
+		);
 	}
 
 	/**
@@ -181,6 +194,23 @@ class WPBT_Extras {
 			},
 			10,
 			2
+		);
+	}
+
+	/**
+	 * Remove auto-installed plugins installed with every beta testing offer.
+	 *
+	 * @return void
+	 */
+	public function remove_auto_installed_plugins() {
+		if ( ! isset( self::$options['remove_auto_installed_plugins'] ) ) {
+			return;
+		}
+		add_action(
+			'init',
+			function () {
+				delete_plugins( array( 'akismet/akismet.php', 'hello.php' ) );
+			}
 		);
 	}
 }